fix login (class in separate module)
