+++ login/models.py
@@ -17,9 +17,6 @@ from django.core.exceptions import PermissionDenied
 from django.db import models
 from django.db.models.signals import post_save
 from django.dispatch import receiver
-from django.contrib.auth.backends import ModelBackend
-from django.db.models import Q
-
 from rest_framework.authtoken.models import Token
 
 import dataedit.models as datamodels
@@ -290,24 +287,3 @@ class UserBackend(object):
 def create_auth_token(sender, instance=None, created=False, **kwargs):
     if created:
         Token.objects.create(user=instance)
-
-
-class ModelBackendWithEmail(ModelBackend):
-    def authenticate(self, request, username=None, password=None, **kwargs):
-        """Change authenticate of ModelBackend
-        
-        Args:
-            request: not used 
-            username(str): username or email from login form
-            password(str): user password from login form        
-        """
-        try: 
-            # find user object based on name OR email
-            user = myuser.objects.get(Q(name=username) | Q(email=username))
-        except models.ObjectDoesNotExist:
-            return None
-
-        if user.check_password(password):
-            return user
-        else:
-            return None
